requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
crewai==0.1.0
python-dotenv==1.0.0
sqlite3
requests==2.31.0